# swagger/swagger.yml
openapi: 3.0.3
info:
  title: Eco-monitoring - Gas analyzers
  version: "1.0"
servers:
  - url: http://localhost:8000
    description: web-app-api
paths:
  /calculate:
    get:
      summary: Calculate Data
      tags:
        - Calculate Data
      operationId: getCalculatedData
      parameters:
        - name: longitude
          in: query
          required: true
          description: Longitude
          schema:
            type: string
          example: 57.99977
        - name: latitude
          in: query
          required: true
          description: Latitude
          schema:
            type: number
          example: 57.99977
        - name: winddirection
          in: query
          required: true
          description: WindDirection
          schema:
            type: number
          example: 270
      responses:
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


  /cities:
    get:
      summary: The method of obtaining a list of cities
      tags:
        - Cities
      operationId: getAllCities
      responses:
        '200':
          description: Getting a list of cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cities"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: The method of creating a city
      tags:
        - Cities
      operationId: createCity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityPost"
      responses:
        '200':
          description: Successful creation of the city
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cities/{id}:
    get:
      summary: The method of obtaining a one city
      tags:
        - Cities
      operationId: getCityById
      parameters:
        - name: id
          in: path
          required: true
          description: City ID
          schema:
            type: string
          example: 13
      responses:
        '200':
          description:  Getting a city by Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: The method of deleting a city
      tags:
        - Cities
      operationId: deleteCityById
      parameters:
        - name: id
          in: path
          required: true
          description: City ID
          schema:
            type: string
          example: 13
      responses:
        '200':
          description: Successful deletion of the city
          content:
            application/json: {}
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


  /gas_analyzers:
    get:
      summary: The method of obtaining a list of gas analyzers
      tags:
        - Gas Analyzers
      operationId: getAllGasAnalyzers
      responses:
        '200':
          description: Getting a list of Gas Analyzers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GasAnalyzers"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: The method of creating a gas analyzer
      tags:
        - Gas Analyzers
      operationId: createGasAnalyzer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GasAnalyzer"
      responses:
        '200':
          description: Successful creation of the gas analyzer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GasAnalyzer"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gas_analyzers/{measurement}:
    get:
      summary: The method of obtaining a gas analyzer
      tags:
        - Gas Analyzers
      operationId: getGasAnalyzerById
      parameters:
        - name: measurement
          in: path
          required: true
          description: Gaz Analyzer Measurement
          schema:
            type: string
          example: gas_analyzer_1
      responses:
        '200':
          description:  Getting a gas analyzer by id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GasAnalyzer"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: The method of deleting a gas analyzer
      tags:
        - Gas Analyzers
      operationId: deleteGasAnalyzerById
      parameters:
        - name: measurement
          in: path
          required: true
          description: Gaz Analyzer Measurement
          schema:
            type: string
          example: gas_analyzer_1
      responses:
        '200':
          description: Successful deletion of the gas analyzer
          content:
            application/json: {}
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gas_analyzers/city/{city_id}:
    get:
      summary: The method of obtaining a list of gas analyzers in city
      tags:
        - Gas Analyzers
      operationId: getGasAnalyzersByCityId
      parameters:
        - name: city_id
          in: path
          required: true
          description: City ID
          schema:
            type: string
          example: 1
      responses:
        '200':
          description:  Getting a list of gas analyzers by city id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GasAnalyzer"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


  /companies:
    get:
      summary: The method of obtaining a list of companies
      tags:
        - Companies
      operationId: getAllCompanies
      responses:
        '200':
          description: Getting a list of companies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Companies"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: The method of creating a company
      tags:
        - Companies
      operationId: createCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyPost"
      responses:
        '200':
          description: Successful creation of the company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /companies/{id}:
    get:
      summary: The method of obtaining a company
      tags:
        - Companies
      operationId: getCompanyById
      parameters:
        - name: id
          in: path
          required: true
          description: Company ID
          schema:
            type: string
          example: 99
      responses:
        '200':
          description:  Getting a company by id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: The method of deleting a company
      tags:
        - Companies
      operationId: deleteCompanyById
      parameters:
        - name: id
          in: path
          required: true
          description: Company ID
          schema:
            type: string
          example: 99
      responses:
        '200':
          description: Successful deletion of the company
          content:
            application/json: {}
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /companies/city/{city_id}:
    get:
      summary: The method of obtaining a list of companies in city
      tags:
        - Companies
      operationId: getCompaniesByCityId
      parameters:
        - name: city_id
          in: path
          required: true
          description: City ID
          schema:
            type: string
          example: 1
      responses:
        '200':
          description:  Getting a list of companies by city id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Companies"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


  /pipes:
    get:
      summary: The method of obtaining a list of pipes
      tags:
        - Pipes
      operationId: getAllPipes
      responses:
        '200':
          description: Getting a list of pipes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipes"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: The method of creating a pipe
      tags:
        - Pipes
      operationId: createPipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pipe"
      responses:
        '200':
          description: Successful creation of the pipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipe"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pipes/{measurement}:
    get:
      summary: The method of obtaining a pipe
      tags:
        - Pipes
      operationId: getPipeById
      parameters:
        - name: measurement
          in: path
          required: true
          description: Pipe Measurement
          schema:
            type: string
          example: pipe_1
      responses:
        '200':
          description:  Getting a pipe by id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipe"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: The method of deleting a pipe
      tags:
        - Pipes
      operationId: deletePipeById
      parameters:
        - name: measurement
          in: path
          required: true
          description: Pipe Measurement
          schema:
            type: string
          example: pipe_1
      responses:
        '200':
          description: Successful deletion of the pipe
          content:
            application/json: {}
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pipes/company/{company_id}:
    get:
      summary: The method of obtaining a list of pipes in company
      tags:
        - Pipes
      operationId: getPipesByCityId
      parameters:
        - name: company_id
          in: path
          required: true
          description: Company ID
          schema:
            type: string
          example: 1
      responses:
        '200':
          description:  Getting a list of pipes by company id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipes"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /guide:
    get:
      summary: The method of obtaining a guide
      tags:
        - Guide
      operationId: getGuide
      responses:
        '200':
          description: Getting a guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guide"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: The method of creating a guide record
      tags:
        - Guide
      operationId: createGuideRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuidePost"
      responses:
        '200':
          description: Successful creation of the guide record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guide"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /guide/{id}:
    delete:
      summary: The method of deleting a guide record
      tags:
        - Guide
      operationId: deleteGuideRecordById
      parameters:
        - name: id
          in: path
          required: true
          description: Guide Record ID
          schema:
            type: string
          example: 1
      responses:
        '200':
          description: Successful deletion of the record
          content:
            application/json: { }
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /influx/gas_analyzer:
    get:
      summary: The method of obtaining a data of gas analyzer
      tags:
        - Influx Data
      operationId: getGasAnalyzerData
      parameters:
        - name: measurement
          in: query
          required: true
          description: Gas Analyzer Measurement
          schema:
            type: string
          example: Gaz_Analyzer_1
        - name: date_from
          in: query
          required: true
          description: Data from
          schema:
            type: string
          example: "2022-01-01 00:00:00"
        - name: date_to
          in: query
          required: false
          description: Data to
          schema:
            type: string
          example: "2022-01-04 00:00:00"
      responses:
        '200':
          description: Getting a list of Gas Analyzer data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GasAnalyzerDatas"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /influx/weather:
    get:
      summary: The method of obtaining a data of gas analyzer
      tags:
        - Influx Data
      operationId: getWeatherData
      parameters:
        - name: measurement
          in: query
          required: true
          description: Weather Measurement
          schema:
            type: string
          example: Perm
        - name: date_from
          in: query
          required: true
          description: Data from
          schema:
            type: string
          example: "2022-01-01 00:00:00"
        - name: date_to
          in: query
          required: false
          description: Data to
          schema:
            type: string
          example: "2022-01-04 00:00:00"
      responses:
        '200':
          description: Getting a list of Gas Analyzer data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherDatas"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /influx/data:
    post:
      summary: The method of creating data of gas analyzer
      tags:
        - Influx Data
      operationId: postInfluxData
      parameters:
        - name: measurement
          in: query
          required: true
          description: Gas Analyzer Measurement
          schema:
            type: string
          example: Gaz_Analyzer_1
        - name: value
          in: query
          required: true
          description: Gas Analyzer Value
          schema:
            type: number
          example: 0.2
      responses:
        '200':
          description: Successful creation of the data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        'default':
          description: Something is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"





components:
  schemas:
    City:
      type: object
      required:
        - name
        - longitude
        - latitude
      properties:
        id:
          type: integer
          example: 13
        name:
          type: string
          example: Perm
        latitude:
          type: number
          example: 58.029682
        longitude:
          type: number
          example: 56.266792

    Guide_Record:
      type: object
      required:
        - substance_name
        - pdk_mr
        - pdk_ss
      properties:
        id:
          type: integer
          example: 13
        substance_name:
          type: string
          example: PM_10
        pdk_mr:
          type: number
          example: 0.5
        pdk_ss:
          type: number
          example: 0.5

    CityPost:
      type: object
      required:
        - name
        - longitude
        - latitude
      properties:
        name:
          type: string
          example: Perm
        latitude:
          type: number
          example: 58.029682
        longitude:
          type: number
          example: 56.266792

    GuidePost:
      type: object
      required:
        - substance_name
        - pdk_mr
        - pdk_ss
      properties:
        substance_name:
          type: string
          example: PM_10
        pdk_mr:
          type: number
          example: 0.5
        pdk_ss:
          type: number
          example: 0.5


    Cities:
      type: array
      items:
        $ref: "#/components/schemas/City"

    Guide:
      type: array
      items:
        $ref: "#/components/schemas/Guide_Record"


    Company:
      type: object
      required:
        - city_id
        - name
        - sanitary_zone_radius
        - longitude
        - latitude
      properties:
        id:
          type: integer
          example: 99
        city_id:
          type: integer
          example: 1
        name:
          type: string
          example: Company-1
        sanitary_zone_radius:
          type: number
          example: 100
        latitude:
          type: number
          example: 58.029682
        longitude:
          type: number
          example: 56.266792
        description:
          type: string
          example: Some description

    CompanyPost:
      type: object
      required:
        - city_id
        - name
        - sanitary_zone_radius
        - longitude
        - latitude
      properties:
        city_id:
          type: integer
          example: 13
        name:
          type: string
          example: Company-1
        sanitary_zone_radius_m:
          type: number
          example: 100
        latitude:
          type: number
          example: 58.029682
        longitude:
          type: number
          example: 56.266792
        description:
          type: string
          example: Some description


    Companies:
      type: array
      items:
        $ref: "#/components/schemas/Company"


    Pipe:
      type: object
      required:
        - measurement
        - company_id
        - longitude
        - latitude
      properties:
        measurement:
          type: string
          example: some_pipe_measurement
        company_id:
          type: integer
          example: 99
        latitude:
          type: number
          example: 58.029682
        longitude:
          type: number
          example: 56.266792


    Pipes:
      type: array
      items:
        $ref: "#/components/schemas/Pipe"


    GasAnalyzer:
      type: object
      required:
        - measurement
        - city_id
        - longitude
        - latitude
        - description
      properties:
        measurement:
          type: string
          example: some_gasanalyzer_measurement
        city_id:
          type: integer
          example: 13
        latitude:
          type: number
          example: 58.029682
        longitude:
          type: number
          example: 56.266792
        description:
          type: string
          example: Some description


    GasAnalyzers:
      type: array
      items:
        $ref: "#/components/schemas/GasAnalyzer"


    GasAnalyzerData:
      type: object
      required:
        - _time
        - _measurement
        - value
      properties:
        _time:
          type: string
          example: "2022-01-02T05:00:18+00:00"
        _measurement:
          type: string
          example: "Gaz_Analyzer_1"
        value:
          type: number
          example: 0.1

    GasAnalyzerDatas:
      type: array
      items:
        $ref: "#/components/schemas/GasAnalyzerData"

    WeatherData:
      type: object
      required:
        - _time
        - _measurement
        - temperature
        - weathercode
        - winddirection
        - windspeed
      properties:
        _time:
          type: string
          example: "2023-03-12T09:41:02.971182+00:00"
        _measurement:
          type: string
          example: "Perm"
        temperature:
          type: number
          example: -4.7
        weathercode:
          type: integer
          example: 3
        winddirection:
          type: number
          example: 153.0
        windspeed:
          type: number
          example: 4.47

    WeatherDatas:
      type: array
      items:
        $ref: "#/components/schemas/WeatherData"


    Success:
     type: object
     required:
      - code
      - message
     properties:
      code:
        type: integer
        example: 200
      message:
        type: string
        example: "success"


    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string